requirements.txt
Flask
ultralytics
opencv-python
google-generativeai
python-dotenv
requests
Pillow
imagehash 